requirements.txt
Django==3.0.3
psycopg2-binary==2.8.4
djangorestframework==3.11.0
httpie==2.0.0